+++ integrations/nextcloud/snappymail/lib/AppInfo/Application.php
@@ -13,8 +13,11 @@ use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
 use OCP\AppFramework\Bootstrap\IBootContext;
+use OCP\App\IAppManager;
+use OCP\EventDispatcher\IEventDispatcher;
+use OCP\IConfig;
 use OCP\IL10N;
-use OCP\IUser;
+use OCP\ISession;
 use OCP\User\Events\PostLoginEvent;
 use OCP\User\Events\BeforeUserLoggedOutEvent;
 use OCA\OIDCLogin\Events\AccessTokenUpdatedEvent;
@@ -47,8 +50,8 @@ class Application extends App implements IBootstrap
 				return new FetchController(
 					$c->query('AppName'),
 					$c->query('Request'),
-					$c->getServer()->getAppManager(),
-					$c->query('ServerContainer')->getConfig(),
+					$c->query(IAppManager::class),
+					$c->query(IConfig::class),
 					$c->query(IL10N::class)
 				);
 			}
@@ -72,12 +75,16 @@ class Application extends App implements IBootstrap
 
 	public function boot(IBootContext $context): void
 	{
-		if (!\is_dir(\rtrim(\trim(\OC::$server->getSystemConfig()->getValue('datadirectory', '')), '\\/') . '/appdata_snappymail')) {
+		$server = $context->getServerContainer();
+		$config = $server->get(IConfig::class);
+		$session = $server->get(ISession::class);
+
+		if (!\is_dir(\rtrim(\trim($config->getSystemValue('datadirectory', '')), '\\/') . '/appdata_snappymail')) {
 			return;
 		}
 
-		$dispatcher = $context->getAppContainer()->query('OCP\EventDispatcher\IEventDispatcher');
-		$dispatcher->addListener(PostLoginEvent::class, function (PostLoginEvent $Event) {
+		$dispatcher = $server->get(IEventDispatcher::class);
+		$dispatcher->addListener(PostLoginEvent::class, function (PostLoginEvent $Event) use ($session) {
 /*
 			$config = \OC::$server->getConfig();
 			// Only store the user's password in the current session if they have
@@ -86,8 +93,8 @@ class Application extends App implements IBootstrap
 			 || $config->getAppValue('snappymail', 'snappymail-autologin-with-email', false)) {
 */
 				$sUID = $Event->getUser()->getUID();
-				\OC::$server->getSession()['snappymail-nc-uid'] = $sUID;
-				\OC::$server->getSession()['snappymail-passphrase'] = SnappyMailHelper::encodePassword($Event->getPassword(), $sUID);
+				$session->set('snappymail-nc-uid', $sUID);
+				$session->set('snappymail-passphrase', SnappyMailHelper::encodePassword($Event->getPassword(), $sUID));
 /*
 			}
 */
@@ -105,13 +112,13 @@ class Application extends App implements IBootstrap
 		// https://github.com/nextcloud/impersonate/pull/180
 		$class = 'OCA\Impersonate\Events\BeginImpersonateEvent';
 		if (\class_exists($class)) {
-			$dispatcher->addListener($class, function ($Event) {
-				\OC::$server->getSession()['snappymail-passphrase'] = '';
+			$dispatcher->addListener($class, function ($Event) use ($session) {
+				$session->set('snappymail-passphrase', '');
 				SnappyMailHelper::loadApp();
 				\RainLoop\Api::Actions()->Logout(true);
 			});
-			$dispatcher->addListener('OCA\Impersonate\Events\EndImpersonateEvent', function ($Event) {
-				\OC::$server->getSession()['snappymail-passphrase'] = '';
+			$dispatcher->addListener('OCA\Impersonate\Events\EndImpersonateEvent', function ($Event) use ($session) {
+				$session->set('snappymail-passphrase', '');
 				SnappyMailHelper::loadApp();
 				\RainLoop\Api::Actions()->Logout(true);
 			});