+++ plugins/nextcloud/index.php
@@ -4,8 +4,8 @@ class NextcloudPlugin extends \RainLoop\Plugins\AbstractPlugin
 {
 	const
 		NAME = 'Nextcloud',
-		VERSION = '2.38.1',
-		RELEASE  = '2024-10-08',
+		VERSION = '2.38.2',
+		RELEASE  = '2026-06-18',
 		CATEGORY = 'Integrations',
 		DESCRIPTION = 'Integrate with Nextcloud v20+',
 		REQUIRED = '2.38.0';
@@ -67,7 +67,7 @@ class NextcloudPlugin extends \RainLoop\Plugins\AbstractPlugin
 
 	public static function IsLoggedIn()
 	{
-		return static::IsIntegrated() && \OC::$server->getUserSession()->isLoggedIn();
+		return static::IsIntegrated() && \OC::$server->get(\OCP\IUserSession::class)->isLoggedIn();
 	}
 
 	public function loginCredentials(string &$sEmail, string &$sLogin, ?string &$sPassword = null) : void
@@ -83,7 +83,7 @@ class NextcloudPlugin extends \RainLoop\Plugins\AbstractPlugin
 
 	public function loginCredentials2(string &$sEmail, ?string &$sPassword = null) : void
 	{
-		$ocUser = \OC::$server->getUserSession()->getUser();
+		$ocUser = \OC::$server->get(\OCP\IUserSession::class)->getUser();
 		$sEmail = $ocUser->getEMailAddress() ?: $ocUser->getPrimaryEMailAddress() ?: $sEmail;
 	}
 
@@ -98,7 +98,7 @@ class NextcloudPlugin extends \RainLoop\Plugins\AbstractPlugin
 		 && \str_starts_with($oSettings->passphrase, 'oidc_login|')
 		) {
 //			$oSettings->passphrase = \OC::$server->getSession()->get('snappymail-passphrase');
-			$oSettings->passphrase = \OC::$server->getSession()->get('oidc_access_token');
+			$oSettings->passphrase = \OC::$server->get(\OCP\ISession::class)->get('oidc_access_token');
 			\array_unshift($oSettings->SASLMechanisms, 'OAUTHBEARER');
 		}
 	}
@@ -213,9 +213,9 @@ class NextcloudPlugin extends \RainLoop\Plugins\AbstractPlugin
 	public function FilterAppData($bAdmin, &$aResult) : void
 	{
 		if (!$bAdmin && \is_array($aResult)) {
-			$ocUser = \OC::$server->getUserSession()->getUser();
+			$ocUser = \OC::$server->get(\OCP\IUserSession::class)->getUser();
 			$sUID = $ocUser->getUID();
-			$oUrlGen = \OC::$server->getURLGenerator();
+			$oUrlGen = \OC::$server->get(\OCP\IURLGenerator::class);
 			$sWebDAV = $oUrlGen->getAbsoluteURL($oUrlGen->linkTo('', 'remote.php') . '/dav');
 //			$sWebDAV = \OCP\Util::linkToRemote('dav');
 			$aResult['Nextcloud'] = [
@@ -225,7 +225,7 @@ class NextcloudPlugin extends \RainLoop\Plugins\AbstractPlugin
 //				'WebDAV_files' => $sWebDAV . '/files/' . $sUID
 			];
 			if (empty($aResult['Auth'])) {
-				$config = \OC::$server->getConfig();
+				$config = \OC::$server->get(\OCP\IConfig::class);
 				$sEmail = '';
 				// Only store the user's password in the current session if they have
 				// enabled auto-login using Nextcloud username or email address.
@@ -269,7 +269,7 @@ class NextcloudPlugin extends \RainLoop\Plugins\AbstractPlugin
 					$aResult['ContactsSync']['User'] = $sUID;
 					$bSave = true;
 				}
-				$pass = \OC::$server->getSession()['snappymail-passphrase'];
+				$pass = \OC::$server->get(\OCP\ISession::class)->get('snappymail-passphrase');
 				if ($pass/* && empty($aResult['ContactsSync']['Password'])*/) {
 					$pass = \SnappyMail\Crypt::DecryptUrlSafe($pass, $sUID);
 					if ($pass) {
@@ -297,8 +297,8 @@ class NextcloudPlugin extends \RainLoop\Plugins\AbstractPlugin
 	{
 		if (!\RainLoop\Api::Config()->Get('webmail', 'allow_languages_on_settings', true)) {
 			$aResultLang = \SnappyMail\L10n::getLanguages($bAdmin);
-			$userId = \OC::$server->getUserSession()->getUser()->getUID();
-			$userLang = \OC::$server->getConfig()->getUserValue($userId, 'core', 'lang', 'en');
+			$userId = \OC::$server->get(\OCP\IUserSession::class)->getUser()->getUID();
+			$userLang = \OC::$server->get(\OCP\IConfig::class)->getUserValue($userId, 'core', 'lang', 'en');
 			$userLang = \strtr($userLang, '_', '-');
 			$sLanguage = $this->determineLocale($userLang, $aResultLang);
 			// Check if $sLanguage is null